Working towards removing a bunch of boilerplate
